DRILL-4576: Add PlannerCallback interface for additional planner initialization.
- Allow a storage plugin to be able to provide additional planner initialization.
- Fix issue in StoragePluginRegistryImpl which deletes manually added plugins from the registry since they don't exist in the PersistentStore.
- Add a simple test to ensure Callbacks are correctly registered and executed.

+++ exec/java-exec/src/main/java/org/apache/drill/exec/store/StoragePluginRegistryImpl.java
@@ -31,6 +31,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 
@@ -80,6 +81,8 @@ public class StoragePluginRegistryImpl implements StoragePluginRegistry {
   private final PersistentStore<StoragePluginConfig> pluginSystemTable;
   private final LogicalPlanPersistence lpPersistence;
   private final ScanResult classpathScan;
+  private final ConcurrentHashMap<String, Integer> manuallyAddedPlugins = new ConcurrentHashMap<>();
+
   private final LoadingCache<StoragePluginConfig, StoragePlugin> ephemeralPlugins;
 
   public StoragePluginRegistryImpl(DrillbitContext context) {
@@ -174,7 +177,9 @@ public class StoragePluginRegistryImpl implements StoragePluginRegistry {
 
       activePlugins.put(INFORMATION_SCHEMA_PLUGIN, new InfoSchemaStoragePlugin(new InfoSchemaConfig(), context,
           INFORMATION_SCHEMA_PLUGIN));
+      manuallyAddedPlugins.put(INFORMATION_SCHEMA_PLUGIN, 0);
       activePlugins.put(SYS_PLUGIN, new SystemTablePlugin(SystemTablePluginConfig.INSTANCE, context, SYS_PLUGIN));
+      manuallyAddedPlugins.put(SYS_PLUGIN, 0);
 
       return activePlugins;
     } catch (IOException e) {
@@ -186,12 +191,15 @@ public class StoragePluginRegistryImpl implements StoragePluginRegistry {
   @Override
   public void addPlugin(String name, StoragePlugin plugin) {
     plugins.put(name, plugin);
+    manuallyAddedPlugins.put(name, 0);
   }
 
   public void deletePlugin(String name) {
     StoragePlugin plugin = plugins.remove(name);
     closePlugin(plugin);
-    pluginSystemTable.delete(name);
+    if(manuallyAddedPlugins.remove(name) == null){
+      pluginSystemTable.delete(name);
+    }
   }
 
   private void closePlugin(StoragePlugin plugin) {
@@ -245,7 +253,7 @@ public class StoragePluginRegistryImpl implements StoragePluginRegistry {
 
   public StoragePlugin getPlugin(String name) throws ExecutionSetupException {
     StoragePlugin plugin = plugins.get(name);
-    if (name.equals(SYS_PLUGIN) || name.equals(INFORMATION_SCHEMA_PLUGIN)) {
+    if (manuallyAddedPlugins.containsKey(name)) {
       return plugin;
     }
 
@@ -354,7 +362,7 @@ public class StoragePluginRegistryImpl implements StoragePluginRegistry {
         }
         // remove those which are no longer in the registry
         for (String pluginName : currentPluginNames) {
-          if (pluginName.equals(SYS_PLUGIN) || pluginName.equals(INFORMATION_SCHEMA_PLUGIN)) {
+          if (manuallyAddedPlugins.containsKey(pluginName)) {
             continue;
           }
           plugins.remove(pluginName);
